+++ main.cpp
@@ -4,6 +4,7 @@
 #include "gcm.h"
 #include "directory.h"
 #include "help.h"
+#include "ram.h"
 
 using namespace std;
 
@@ -45,5 +46,29 @@ int main()
 
 //    string n = "1.txt";
 //    cout << statictic(n);
+
+//    FileOpenList f;
+//    int j = f.addItem(0, 1, 2, 3);
+//    f.show();
+//    cout << f.size() << endl;
+//    int i = f.addItem(100, 3, 4, 5);
+//    f.addLink(i);
+//    f.show();
+//    cout << f.size() << endl;
+//    f.deleteLink(j);
+//    f.show();
+//    cout << f.size() << endl;
+
+//    UserOpenList u("cyndsb");
+//    cout << u.searchFreeItem() << endl;
+//    u.addItem(2, 4);
+//    u.addItem(88, 99);
+//    cout << u.searchId(3) << endl;
+//    cout << u.searchId(2) << endl;
+//    cout << u.getFileId(2) << endl;
+//    u.deleteItem(88);
+//    cout << u.getFileId(2) << endl;
+//    cout << u.searchFreeItem() << endl;
+//    cout << u.searchId(2) << endl;
 	return 0;
 }
